🐴 `spatial`: changed `distance_matrix` parameter default from int to float

+++ scipy-stubs/spatial/_kdtree.pyi
@@ -117,8 +117,10 @@ def minkowski_distance(x: onp.ToComplexND, y: onp.ToComplexND, p: int = 2) -> on
 
 #
 @overload
-def distance_matrix(x: onp.ToFloatND, y: onp.ToFloatND, p: int = 2, threshold: int = 1_000_000) -> onp.Array2D[np.float64]: ...
+def distance_matrix(
+    x: onp.ToFloatND, y: onp.ToFloatND, p: float = 2.0, threshold: int = 1_000_000
+) -> onp.Array2D[np.float64]: ...
 @overload
 def distance_matrix(
-    x: onp.ToComplexND, y: onp.ToComplexND, p: int = 2, threshold: int = 1_000_000
+    x: onp.ToComplexND, y: onp.ToComplexND, p: float = 2.0, threshold: int = 1_000_000
 ) -> onp.Array2D[np.float64 | np.complex128]: ...
